use is_same_v in wrap::narrow to determine action
use std::is_same_v to check if input types are the same and therefore
whether the narrow operation is actually neccessary. This allows us to
use the most efficient code consistent with security when the types
vary between x64 and x32 versions.

replace all instances where the input types are now the same regardless