fix the redundant initialization in FRM_HEAD_OUT
angleOrClipData does not need to be initialized since FANGCLPOUT has
a constructor. Doing this does mean that the FRM_HEAD_OUT constructor
cannot be constexpr unless all of the relevant members are also
constexpr. Specifically, the constructors for FANGCLPOUT, FLENCNTOUT,
TXTR_INFO & SATINANGLEOUT are now declared as constexpr. Doing this
also allowed the FANGCLPOUT constructor to be defaulted.