fix the redundant initialization in FRM_HEAD_OUT
fillInfo does not need to be initialized since TF_INFO_OUT has
a constructor. Doing this does mean that the FRM_HEAD_OUT constructor
cannot be constexpr unless all of the relevant members are also
constexpr. Specifically, the constructor for FTHR_INFO & TF_INFO_OUT
are now declared as constexpr. The TF_INFO_OUT constructor cannot be
defaulted though, so a simple constructor is also declared