optimize loop in ritbal
Remove the iOutput variable. iOutput was used to track the records
output but the size of the balaradStitch vector gives us that
anyway. Use wrapper function to calculate the file write size
Convert the loop to a range-based for loop over StitchBuffer. Use
iStitch to track the current position in StitchBuffer.