Replace longer and shorter with nextSortedStitch
longer searched for the next longest stitch after the current stitch.
shorter found the next shortest stitch after the current stitch. Both
of these suffered from the fact that the next stitch found in either
direction may be a stitch already found previously. So, instead we
sort all the stitches and traverse the sorted vector in the direction
requested.