don't set the fill type in deltx
This fixes the problem where isTexture would always return false.
This meant that the texture deletion would never occur and in fact,
the body of deltx was optimized out in the release version. Every
call to deltx is succeeded by code to set the fill type, so this does
not break anything else.